- Fixed the ps file generation to show more points with the set color

+++ geometry/src/main/headers/Geometry.h
@@ -5,7 +5,7 @@
 #include <cassert>
 namespace Geometry {
 
-	typedef int CoordinateType;
+	typedef float CoordinateType;
 	typedef double DistanceType;
 	typedef double AreaType;
 	typedef unsigned int SizeType;
